Standardize defines and file formatting
Updates all the files to use the same #define statements, with matching names. Also updates the setup() function to use a common pattern for the FastLED.addLeds() function call. Also updates the README with more explicit instructions for customizing and using these sample.This should make it easier

These changes should make it easier to use these samples. Novices will find a common pattern, which will be easier to follow. Experienced users will be able to more quickly use each of the samples.